LayoutTests/platform/win/http/tests/misc/webtiming-slow-load-expected.txt

Verifies that requestStart and responseStart are available before the main document has finished loading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL timing.requestStart should be non-zero. Was 0
FAIL timing.responseStart should be non-zero. Was 0
FAIL timing.responseEnd should be non-zero. Was 0
PASS successfullyParsed is true

TEST COMPLETE
